+++ src/test/java/io/pivotal/cla/service/github/MylynGitHubApiITests.java
@@ -32,7 +32,6 @@ import org.springframework.core.serializer.support.SerializingConverter;
 
 import io.pivotal.cla.config.ClaOAuthConfig;
 import io.pivotal.cla.config.OAuthClientCredentials;
-import io.pivotal.cla.data.User;
 import okhttp3.mockwebserver.EnqueueRequests;
 import okhttp3.mockwebserver.EnqueueResourcesMockWebServer;
 import okhttp3.mockwebserver.RecordedRequest;
@@ -83,184 +82,6 @@ public class MylynGitHubApiITests {
 		assertThat(request.getMethod()).isEqualTo("GET");
 		assertThat(request.getPath()).isEqualTo("/api/v3/user/repos?per_page=100&page=1");
 		assertThat(request.getHeader("Authorization")).isEqualTo("token " + accessToken);
-	}
-
-	@Test
-	@EnqueueRequests({"getAccessToken",
-		"getEmailsNotPivotal",
-		"getUserRwinch"})
-	public void getCurrentUserAdminRequestedButNotAdmin() throws Exception {
-		OAuthAccessTokenParams oauthParams = new OAuthAccessTokenParams();
-		oauthParams.setCallbackUrl("https://example.com/oauth/callback");
-		oauthParams.setCode("code-123");
-		oauthParams.setState("state-456");
-
-		CurrentUserRequest userRequest = new CurrentUserRequest();
-		userRequest.setOauthParams(oauthParams);
-		userRequest.setRequestAdminAccess(true);
-
-		User user = service.getCurrentUser(userRequest);
-
-		assertThat(user.getAccessToken()).isEqualTo("access-token-123");
-		assertThat(user.getAvatarUrl()).isEqualTo("https://avatars.githubusercontent.com/u/362503?v=3");
-		assertThat(user.getEmails()).containsOnly("rob@example.com");
-		assertThat(user.getGitHubLogin()).isEqualTo("rwinch");
-		assertThat(user.getFullName()).isEqualTo("Rob Winch");
-		assertThat(user.isAdminAccessRequested()).isTrue();
-		assertThat(user.isAdmin()).isFalse();
-
-		RecordedRequest request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("POST");
-		assertThat(request.getPath())
-				.isEqualTo("/login/oauth/access_token");
-		assertThat(request.getBody().readUtf8())
-			.isEqualTo("{\"code\":\"code-123\",\"client_secret\":\"client-secret\",\"state\":\"state-456\",\"client_id\":\"client-id\",\"redirect_url\":\"https://example.com/oauth/callback\"}");
-
-		request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("GET");
-		assertThat(request.getPath())
-				.isEqualTo("/api/v3/user/emails?per_page=100&page=1");
-		assertThat(request.getHeader("Authorization")).isEqualTo("token " + user.getAccessToken());
-
-		request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("GET");
-		assertThat(request.getPath())
-				.isEqualTo("/api/v3/user");
-		assertThat(request.getHeader("Authorization")).isEqualTo("token " + user.getAccessToken());
-	}
-
-	@Test
-	@EnqueueRequests({"getAccessToken",
-		"getEmailsPivotal",
-		"getUserRwinch",
-		"getTeamMembers"})
-	public void getCurrentUserAdminAndClaAuthor() throws Exception {
-		OAuthAccessTokenParams oauthParams = new OAuthAccessTokenParams();
-		oauthParams.setCallbackUrl("https://example.com/oauth/callback");
-		oauthParams.setCode("code-123");
-		oauthParams.setState("state-456");
-		CurrentUserRequest userRequest = new CurrentUserRequest();
-		userRequest.setOauthParams(oauthParams);
-		userRequest.setRequestAdminAccess(true);
-
-		User user = service.getCurrentUser(userRequest);
-
-		assertThat(user.getAccessToken()).isEqualTo("access-token-123");
-		assertThat(user.getAvatarUrl()).isEqualTo("https://avatars.githubusercontent.com/u/362503?v=3");
-		assertThat(user.getEmails()).containsOnly("rob@example.com","rob@pivotal.io");
-		assertThat(user.getGitHubLogin()).isEqualTo("rwinch");
-		assertThat(user.getFullName()).isEqualTo("Rob Winch");
-		assertThat(user.isAdminAccessRequested()).isTrue();
-		assertThat(user.isAdmin()).isTrue();
-		assertThat(user.isClaAuthor()).isTrue();
-
-		RecordedRequest request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("POST");
-		assertThat(request.getPath())
-				.isEqualTo("/login/oauth/access_token");
-		assertThat(request.getBody().readUtf8())
-			.isEqualTo("{\"code\":\"code-123\",\"client_secret\":\"client-secret\",\"state\":\"state-456\",\"client_id\":\"client-id\",\"redirect_url\":\"https://example.com/oauth/callback\"}");
-
-		request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("GET");
-		assertThat(request.getPath())
-				.isEqualTo("/api/v3/user/emails?per_page=100&page=1");
-		assertThat(request.getHeader("Authorization")).isEqualTo("token " + user.getAccessToken());
-
-		request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("GET");
-		assertThat(request.getPath())
-				.isEqualTo("/api/v3/user");
-		assertThat(request.getHeader("Authorization")).isEqualTo("token " + user.getAccessToken());
-	}
-
-	@Test
-	@EnqueueRequests({"getAccessToken",
-		"getEmailsPivotal",
-		"getUserRwinch",
-		"getTeamMembersNotFound"})
-	public void getCurrentUserAdminAndNotClaAuthor() throws Exception {
-		OAuthAccessTokenParams oauthParams = new OAuthAccessTokenParams();
-		oauthParams.setCallbackUrl("https://example.com/oauth/callback");
-		oauthParams.setCode("code-123");
-		oauthParams.setState("state-456");
-		CurrentUserRequest userRequest = new CurrentUserRequest();
-		userRequest.setOauthParams(oauthParams);
-		userRequest.setRequestAdminAccess(true);
-
-		User user = service.getCurrentUser(userRequest);
-
-		assertThat(user.getAccessToken()).isEqualTo("access-token-123");
-		assertThat(user.getAvatarUrl()).isEqualTo("https://avatars.githubusercontent.com/u/362503?v=3");
-		assertThat(user.getEmails()).containsOnly("rob@example.com","rob@pivotal.io");
-		assertThat(user.getGitHubLogin()).isEqualTo("rwinch");
-		assertThat(user.getFullName()).isEqualTo("Rob Winch");
-		assertThat(user.isAdminAccessRequested()).isTrue();
-		assertThat(user.isAdmin()).isTrue();
-		assertThat(user.isClaAuthor()).isFalse();
-
-		RecordedRequest request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("POST");
-		assertThat(request.getPath())
-				.isEqualTo("/login/oauth/access_token");
-		assertThat(request.getBody().readUtf8())
-			.isEqualTo("{\"code\":\"code-123\",\"client_secret\":\"client-secret\",\"state\":\"state-456\",\"client_id\":\"client-id\",\"redirect_url\":\"https://example.com/oauth/callback\"}");
-
-		request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("GET");
-		assertThat(request.getPath())
-				.isEqualTo("/api/v3/user/emails?per_page=100&page=1");
-		assertThat(request.getHeader("Authorization")).isEqualTo("token " + user.getAccessToken());
-
-		request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("GET");
-		assertThat(request.getPath())
-				.isEqualTo("/api/v3/user");
-		assertThat(request.getHeader("Authorization")).isEqualTo("token " + user.getAccessToken());
-	}
-
-
-	@Test
-	@EnqueueRequests({"getAccessToken",
-		"getEmailsNotPivotal",
-		"getUserRwinch"})
-	public void getCurrentUserSigning() throws Exception {
-		OAuthAccessTokenParams oauthParams = new OAuthAccessTokenParams();
-		oauthParams.setCallbackUrl("https://example.com/oauth/callback");
-		oauthParams.setCode("code-123");
-		oauthParams.setState("state-456");
-		CurrentUserRequest userRequest = new CurrentUserRequest();
-		userRequest.setOauthParams(oauthParams);
-		userRequest.setRequestAdminAccess(false);
-
-		User user = service.getCurrentUser(userRequest);
-
-		assertThat(user.getAccessToken()).isEqualTo("access-token-123");
-		assertThat(user.getAvatarUrl()).isEqualTo("https://avatars.githubusercontent.com/u/362503?v=3");
-		assertThat(user.getEmails()).containsOnly("rob@example.com");
-		assertThat(user.getGitHubLogin()).isEqualTo("rwinch");
-		assertThat(user.getFullName()).isEqualTo("Rob Winch");
-		assertThat(user.isAdminAccessRequested()).isFalse();
-		assertThat(user.isAdmin()).isFalse();
-
-		RecordedRequest request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("POST");
-		assertThat(request.getPath())
-				.isEqualTo("/login/oauth/access_token");
-		assertThat(request.getBody().readUtf8())
-			.isEqualTo("{\"code\":\"code-123\",\"client_secret\":\"client-secret\",\"state\":\"state-456\",\"client_id\":\"client-id\",\"redirect_url\":\"https://example.com/oauth/callback\"}");
-
-		request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("GET");
-		assertThat(request.getPath())
-				.isEqualTo("/api/v3/user/emails?per_page=100&page=1");
-		assertThat(request.getHeader("Authorization")).isEqualTo("token " + user.getAccessToken());
-
-		request = server.getServer().takeRequest();
-		assertThat(request.getMethod()).isEqualTo("GET");
-		assertThat(request.getPath())
-				.isEqualTo("/api/v3/user");
-		assertThat(request.getHeader("Authorization")).isEqualTo("token " + user.getAccessToken());
 	}
 
 	@Test
