--- src/test/resources/io/pivotal/cla/service/github/MylynGitHubApiITests_okhttp3/getUserRwinch
@@ -1,69 +1,0 @@
-HTTP/1.1 200 OK
-Access-Control-Allow-Origin: *
-Access-Control-Expose-Headers: ETag, Link, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset, X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval
-Cache-Control: private, max-age=60, s-maxage=60
-Content-Security-Policy: default-src 'none'
-Content-Type: application/json; charset=utf-8
-Date: Sat, 05 Mar 2016 02:00:39 GMT
-ETag: W/"e5665150505598abe7d231900bcc652d"
-Last-Modified: Mon, 29 Feb 2016 19:32:24 GMT
-Server: GitHub.com
-Status: 200 OK
-Strict-Transport-Security: max-age=31536000; includeSubdomains; preload
-Vary: Accept, Authorization, Cookie, X-GitHub-OTP
-Vary: Accept-Encoding
-X-Accepted-OAuth-Scopes:
-X-Content-Type-Options: nosniff
-X-Frame-Options: deny
-X-GitHub-Media-Type: github.v3
-X-GitHub-Request-Id: 4ADEC8ED:1D92B:4F6F06:56DA3DC7
-X-OAuth-Scopes: admin:org, admin:org_hook, admin:public_key, admin:repo_hook, delete_repo, gist, notifications, repo, user
-X-RateLimit-Limit: 5000
-X-RateLimit-Remaining: 4998
-X-RateLimit-Reset: 1457145439
-X-Served-By: a51acaae89a7607fd7ee967627be18e4
-X-XSS-Protection: 1; mode=block
-
-{
-    "avatar_url": "https://avatars.githubusercontent.com/u/362503?v=3",
-    "bio": null,
-    "blog": "http://spring.io/team/rwinch",
-    "collaborators": 0,
-    "company": "Pivotal",
-    "created_at": "2010-08-12T18:08:04Z",
-    "disk_usage": 99308,
-    "email": null,
-    "events_url": "https://api.github.com/users/rwinch/events{/privacy}",
-    "followers": 282,
-    "followers_url": "https://api.github.com/users/rwinch/followers",
-    "following": 0,
-    "following_url": "https://api.github.com/users/rwinch/following{/other_user}",
-    "gists_url": "https://api.github.com/users/rwinch/gists{/gist_id}",
-    "gravatar_id": "",
-    "hireable": null,
-    "html_url": "https://github.com/rwinch",
-    "id": 362503,
-    "location": null,
-    "login": "rwinch",
-    "name": "Rob Winch",
-    "organizations_url": "https://api.github.com/users/rwinch/orgs",
-    "owned_private_repos": 0,
-    "plan": {
-        "collaborators": 0,
-        "name": "free",
-        "private_repos": 0,
-        "space": 976562499
-    },
-    "private_gists": 17,
-    "public_gists": 13,
-    "public_repos": 156,
-    "received_events_url": "https://api.github.com/users/rwinch/received_events",
-    "repos_url": "https://api.github.com/users/rwinch/repos",
-    "site_admin": false,
-    "starred_url": "https://api.github.com/users/rwinch/starred{/owner}{/repo}",
-    "subscriptions_url": "https://api.github.com/users/rwinch/subscriptions",
-    "total_private_repos": 0,
-    "type": "User",
-    "updated_at": "2016-02-29T19:32:24Z",
-    "url": "https://api.github.com/users/rwinch"
-}
